created separate routes
